+++ lib/elixir/test/elixir/dict_test.exs
@@ -1,4 +1,4 @@
-Code.require_file "../test_helper.exs", __FILE__
+Code.require_file "test_helper.exs", __DIR__
 
 defmodule DictTest.Common do
   defmacro __using__(module) do