+++ lib/elixir/test/elixir/kernel/cli_test.exs
@@ -1,4 +1,4 @@
-Code.require_file "../../test_helper.exs", __FILE__
+Code.require_file "../test_helper.exs", __DIR__
 
 defmodule Kernel.CLI.InitTest do
   use ExUnit.Case, async: true