+++ lib/mix/test/mix/tasks/run_test.exs
@@ -1,4 +1,4 @@
-Code.require_file "../../../test_helper.exs", __FILE__
+Code.require_file "../../test_helper.exs", __DIR__
 
 defmodule Mix.Tasks.RunTest do
   use MixTest.Case